+++ Cube.h
@@ -1,14 +1,16 @@
-//    * Edge indexes.
+//    * Edge indices.
 //    *
 //    * 0  1  2  3  4  5  6  7  8  9  10 11
 //    * UB UR UF UL FR FL BL BR DF DL DB DR
 //    * BO BY BR BW RY RW OW OY GR GW GO GY
 //    *
-//    * Corner indexes.
+//    * Corner indices.
 //    *
 //    * 0   1   2   3   4   5   6   7
 //    * ULB URB URF ULF DLF DLB DRB DRF
 //    * BWO BOY BYR BRW GWR GOW GYO GRY
+
+// Cube is configured in orientation with blue up, red front
 
 
 #pragma once
@@ -35,6 +37,6 @@ private:
     Cubie edges[12];
     Cubie corners[8];
 
-    void updateCornerOrientation(uint8_t corner_index, uint change);
+    void updateCornerOrientation(uint8_t corner_index, uint8_t change);
 
 };
